% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{powiaty}
\alias{powiaty}
\title{Mapowanie kodów TERYT powiatów na podregiony statystyczne (wg NUTS
2016)}
\format{
Ramka danych z 7 kolumnami i 380 wierszami.
}
\usage{
powiaty
}
\description{
Ramka danych z sześciocyfrowymi (lub pięciocyfrowymi) kodami
TERYT powiatów oraz nazwami powiatów, podregionów statystycznych i województw
w mianowniku, do których poszczególne powiaty są przypisane oraz tych samych
nazw w dopełniaczu.
}
\details{
Ramka danych służy do mapowania nazw jednostek podziału
terytorialnego w szablonach pakietu.
}
\keyword{datasets}
